==> notebooks/03-venue-classifier.ipynb <==
# BERT Venue Classfier Training

Now that we have a labeled venue dataset, we need a way to create classification scores for each persona for each venue. We can use an 8 head classification model for this task. By taking BERT, and feeding the business name, categories, feature list and one sentence description (or some subset of these elements) into the model as input, we can use the persona labels provided by GPT to fine-tune a BERT checkpoint on providing relationship scores between a venue and our set of personas.

1. **Dataset Creation**

We will want to start by creating a labeled dataset that can be fed directly to BERT for a fine-tuning job. This will allow us to setup a function for transforming a row into an input prompt, which we will be able to modify to test different results of the BERT encoder model.

2. **Fine-Tuning**

We will fine-tune the model using the labelled dataset. This should utilize hugginface and the `AutoModelForSequenceClassfication` class, which will wrap our BERT checkpoint and tokenizer.

3. **Evaluation**

After the evaluation is complete, we will assess the model's perfromance on our test dataset. This will allow us to see how well the model distributes the scores for the 8 labels for each venue. We can modify the input string if necessary based on these results.

import json

import pandas as pd
from dotenv import load_dotenv

load_dotenv()


### Dataset Creation

For starters, we will setup a pipeline to take our Yelp location data and create rows that can be used to fine-tune our classifier model.


with open("../data/venues/yelp.json", "r") as f:
    location_data = json.load(f)

personas = ["socialButterfly", "culinaryExplorer", "beautyFashionAficionado", "familyOrientedIndividual", "artCultureEnthusiast", "wellnessSelfCareAdvocate", "adventurerExplorer", "ecoConsciousConsumer"]

dataset = []
for loc in location_data:
    labels = {persona: 1 if persona in loc['personas'] else 0 for persona in personas}
    data = {
        "id": loc["id"],
        "biz_name": loc["name"],
        "categories": ', '.join([cat['title'] for cat in loc["categories"]]),
        "biz_features": loc["biz_features"],
        'summary': loc["business_summary"],
    }
    row = {**data, **labels}
    dataset.append(row)

df = pd.DataFrame(dataset)

def format_input(row):
    return (
        f"Name: {row.biz_name}\n"
        f"Categories: {row.categories}\n"
        f"Biz Features: {row.biz_features}\n"
        f"Summary: {row.summary}\n"
    )

df['input'] = df.apply(format_input, axis=1)

# inputs = tokenizer(df['input'].tolist(), return_tensors='pt', padding=True, truncation=True, max_length=512)
# labels = torch.tensor(df[personas].values, dtype=torch.float)

# dataset = TensorDataset(inputs['input_ids'], inputs['attention_mask'], labels)
# train_size = int(0.9 * len(dataset))
# val_size = len(dataset) - train_size

# train_dataset, val_dataset = random_split(dataset, [train_size, val_size])

# train_dataloader = DataLoader(train_dataset, batch_size=16, shuffle=True)
# val_dataloader = DataLoader(val_dataset, batch_size=16)


### Fine-Tuning the Model

from transformers import AutoModelForSequenceClassification, AutoTokenizer

bert_chkpt = "distilbert-base-uncased"

model = AutoModelForSequenceClassification.from_pretrained(bert_chkpt, num_labels=len(personas), problem_type="multi_label_classification")
tokenizer = AutoTokenizer.from_pretrained(bert_chkpt)

import torch
import numpy as np
from datasets import Dataset, DatasetDict

inputs = tokenizer(df.input.tolist(), return_tensors='pt', padding=True, truncation=True, max_length=512)
labels = torch.tensor(df[personas].values, dtype=torch.float)

indicies = np.arange(len(labels))

np.random.seed(100)
np.random.shuffle(indicies)

train_indicies = indicies[:800]
val_indicies = indicies[800:1000]
test_indicies = indicies[1000:]

train_dataset = Dataset.from_dict({'input_ids': inputs['input_ids'][train_indicies], 'attention_mask': inputs['attention_mask'][train_indicies], 'labels': labels[train_indicies]})
val_dataset = Dataset.from_dict({'input_ids': inputs['input_ids'][val_indicies], 'attention_mask': inputs['attention_mask'][val_indicies], 'labels': labels[val_indicies]})
test_dataset = Dataset.from_dict({'input_ids': inputs['input_ids'][test_indicies], 'attention_mask': inputs['attention_mask'][test_indicies], 'labels': labels[test_indicies]})

dataset_dict = DatasetDict({
    'train': train_dataset,
    'val': val_dataset,
    'test': test_dataset
})


from transformers import Trainer, TrainingArguments

training_args = TrainingArguments(
    output_dir='./results',          # output directory
    num_train_epochs=3,              # total number of training epochs
    per_device_train_batch_size=16,  # batch size per device during training
    per_device_eval_batch_size=64,   # batch size for evaluation
    warmup_steps=500,                # number of warmup steps for learning rate scheduler
    weight_decay=0.01,               # strength of weight decay
    logging_dir='./logs',            # directory for storing logs
)

trainer = Trainer(
    model=model,                            # the instantiated 🤗 Transformers model to be trained
    args=training_args,                     # training arguments, defined above
    train_dataset=dataset_dict['train'],    # training dataset
    eval_dataset=dataset_dict['val']        # evaluation dataset
)

result = trainer.train()

import torch
import numpy as np

label_map = {i: label for i, label in enumerate(personas)}
predictions = trainer.predict(dataset_dict['test'])
predictions_tensor = torch.tensor(predictions.predictions)

predictions = torch.nn.functional.softmax(predictions_tensor, dim=-1)

labels_df = pd.DataFrame(predictions, columns=personas)
labels_df.describe()

trainer.save_model("../models/bert-yelp")
