==> notebooks/04-venue-entity-creation.ipynb <==
# Venue Entity Creation

Now that we have completed the process of filling out the entire pipeline of for venue processing, it is time to create the Venue entities and store them in the database. We will need to create two types of entities for each venue: a **vectorstore document** and a **graph node**. These entities will be able to be used in conjunction to first search the vecotr database to find a venue based on the user's query, and then the graph node will be used to order those results by there relationship to the user's mood board.

**Vectorstore Documents**:

This is where the majority of the metadata for each venue is stored. These documents should carry all of the metadata that we are stroing for each venue, the graph will just be used for relationships with social media posts.

**Graph Nodes**: 

The graph nodes will be created using the BERT classifier. We will classify each venue and create relationships with each persona type, where the weight is the output score of BERT's label classification for that venue, after being passed through a softmax layer.

### Vectorstore Documents

We want to start by creating the vectorstore documents for each venue, which will allow us to search and filter for venues based on specific criteria and a embeded query.

import json
from dotenv import load_dotenv

load_dotenv()

with open("../data/venues/yelp.json", "r") as f:
    location_data = json.load(f)

vector_data = []
for loc in location_data:
    vector_data.append({
        "id": loc["id"],
        "name": loc["name"],
        "description": loc["business_summary"],
        "city": loc['city_code'],
        "hours": loc["hours"],
        "category": loc['category']
    })

from openai import OpenAI
from typing import List

def embed_batch(batch: List[str]):
    client = OpenAI()
    response = client.embeddings.create(
        input=batch,
        model="text-embedding-ada-002"
    )
    return [datum.embedding for datum in response.data]

descriptions = [datum["description"] for datum in vector_data]
embeddings = embed_batch(descriptions)
for row, embedding in zip(vector_data, embeddings):
    row["embedding"] = embedding

# Finally, we need to transform our vector data into a format that is consumeable by
# the pinecone API

pinecone_data = []
for row in vector_data:
    pinecone_data.append({
        "id": row["id"],
        "values": row["embedding"],
        "metadata": {
            "name": row["name"],
            "description": row["description"],
            "city": row["city"],
            "category": row["category"]
        }
    })

with open("../data/venues/vectorstore_docs.json", "w") as f:
    json.dump(pinecone_data, f)

### Graph Nodes

Now that we have our vectorstore docs setup, we need to use our model at `../models/bert-yelp` to classfiy persona labels on each venue. We will then use those persona labels create relationships between each Venue entity and Persona entity, with a weight that is determined by the softmax adjusted score of the label.

from transformers import AutoTokenizer, AutoModelForSequenceClassification

model_tokenizer = "distilbert-base-uncased"
model_path = "../models/bert-yelp"
tokenizer = AutoTokenizer.from_pretrained(model_tokenizer)
model = AutoModelForSequenceClassification.from_pretrained(model_path)

def format_input(row):
    return (
        f"Name: {row.biz_name}\n"
        f"Categories: {row.categories}\n"
        f"Biz Features: {row.biz_features}\n"
        f"Summary: {row.summary}\n"
    )


import pandas as pd

graph_data = []
for loc in location_data:
    graph_data.append({
        "id": loc["id"],
        "city": loc["city_code"],
        "category": loc["category"],
        "biz_name": loc["name"],
        "categories": ', '.join([cat['title'] for cat in loc["categories"]]),
        "biz_features": loc["biz_features"],
        'summary': loc["business_summary"],
        'url': loc["url"],
        'thumbnail_url': loc['image_url']
    })

df = pd.DataFrame(graph_data)

df['input'] = df.apply(format_input, axis=1)

inputs = tokenizer(df.input.tolist(), return_tensors='pt', padding=True, truncation=True, max_length=512)

df['input_ids'] = inputs['input_ids'].tolist()
df['attention_mask'] = inputs['attention_mask'].tolist()

import torch

with torch.no_grad():
    model.eval()
    inputs = {
        'input_ids': torch.tensor(df['input_ids'].tolist()),
        'attention_mask': torch.tensor(df['attention_mask'].tolist())
    }
    outputs = model(**inputs)
    predictions = torch.nn.functional.softmax(outputs.logits, dim=-1)
    df['predictions'] = predictions.tolist()

personas = ["socialButterfly", "culinaryExplorer", "beautyFashionAficionado", "familyOrientedIndividual", "artCultureEnthusiast", "wellnessSelfCareAdvocate", "adventurerExplorer", "ecoConsciousConsumer"]
for i, persona in enumerate(personas):
    df[persona] = df['predictions'].apply(lambda x: x[i])
df.head()

# Now, we want to create an array of objects that can be used to create cypher statments.

cypher_data = []
for _, row in df.iterrows():
    data = {
        "venue": {
            "id": row['id'],
            "name": row['biz_name'],
            "city": row['city'],
            "category": row['category'],
            "url": row['url'],
            "thumbnail_url": row['thumbnail_url'],
        },
        "personas": {
            "socialButterfly": row['socialButterfly'],
            "culinaryExplorer": row['culinaryExplorer'],
            "beautyFashionAficionado": row['beautyFashionAficionado'],
            "familyOrientedIndividual": row['familyOrientedIndividual'],
            "artCultureEnthusiast": row['artCultureEnthusiast'],
            "wellnessSelfCareAdvocate": row['wellnessSelfCareAdvocate'],
            "adventurerExplorer": row['adventurerExplorer'],
            "ecoConsciousConsumer": row['ecoConsciousConsumer'],
        }
    }
    cypher_data.append(data)

with open("../data/venues/cypher_data.json", "w") as f:
    json.dump(cypher_data, f)
